PrismaVeil v3: complete protocol redesign with anti-censorship suite
Redesign the PrismaVeil wire protocol from v2 (4-step, 2 RTT) to v3
(2-step, 1 RTT) with 0-RTT session resumption via ticket system.
Implement all 10 planned phases across the full stack.

Protocol v3:
- 2-step handshake (ClientInit -> ServerInit), 14 command types,
  2-byte flags, challenge-response verification
- Session ticket system with bloom filter anti-replay for 0-RTT
- 2-phase BLAKE3 key derivation (preliminary + final session key)
- Backward compatible: server accepts v1/v2/v3 clients

Transports:
- Add WebSocket, gRPC, XHTTP (packet-up/stream-up/stream-one)
- XMUX connection pooling with configurable concurrency
- CDN-compatible reverse proxy listener

Anti-censorship:
- Salamander UDP obfuscation (BLAKE3-derived XOR keystream over
  quinn::AsyncUdpSocket, cached key for per-packet performance)
- HTTP/3 masquerade via h3/h3-quinn (serves real websites to browsers,
  PrismaVeil clients distinguished by ALPN negotiation)
- Port hopping (HMAC-based deterministic UDP port rotation)
- TLS-on-TCP camouflage with decoy fallback

Performance:
- 3 congestion control modes: Brutal (fixed-rate), BBR, Adaptive
  (auto-detects throttling), as quinn::congestion::Controller impls
- PrismaUDP relay via CMD_UDP_ASSOCIATE/CMD_UDP_DATA with QUIC
  DATAGRAM extension for unreliable low-latency delivery
- FEC (Reed-Solomon erasure coding) for UDP flows with auto-eviction

TUN mode:
- System-wide transparent proxy via virtual network interface
- Windows (Wintun), Linux (/dev/net/tun), macOS (utun)
- smoltcp userspace TCP/IP stack for TCP stream extraction
- DNS interception on port 53

DNS & routing:
- 4 DNS modes: direct, smart (domain blocklist), fake IP (zero leaks
  via 198.18.0.0/15 LRU pool), tunnel (all DNS through proxy)
- Rule-based routing: domain/suffix/keyword/IP-CIDR/port rules
  with proxy/direct/block actions

Operations:
- Per-client bandwidth limits and traffic quotas (daily/weekly/monthly)
- Built-in speed test (CMD_SPEED_TEST)
- CLI commands: init, validate, status, speed-test
- Dashboard: client edit page, auth context provider

Code quality:
- VecDeque for O(1) queue operations (packet queues, LRU eviction)
- Consolidated mutex acquisitions to reduce async contention
- Proper backpressure via reserve_owned() in XHTTP streams
- Shared HTTP client in H3 masquerade (avoids per-request allocation)

Docs & tests:
- Updated README with v3 features, config refs, protocol section,
  project structure (181 tests, up from 58)
- Updated protocol docs, config docs, anti-detection docs
- 72 modified files, 45 new files, ~5800 lines added

+++ prisma-core/tests/protocol_proptest.rs
@@ -37,7 +37,7 @@ proptest! {
     #[test]
     fn test_data_frame_round_trip(
         cmd in arb_command(),
-        flags in any::<u8>(),
+        flags in any::<u16>(),
         stream_id in any::<u32>(),
     ) {
         let frame = DataFrame {
